requirements.txt
setuptools
wheel
pylint
pylint-django
flake8
pydocstyle
